refac: frames.pyにまとまっていたframeクラスを別々のファイルで管理させる
