+++ gui/frames/__init__.py
@@ -1,0 +1,6 @@
+from .crops_frame import CropsFrame
+from .inference_model_frame import InferenceModelFrame
+from .option_frame import OptionFrame
+from .model_selection_frame import ModelSelectionFrame
+
+__all__ = ["CropsFrame", "InferenceModelFrame", "OptionFrame", "ModelSelectionFrame"]
